+++ src/Tools/Metadata/AnalyzerOptionMetadata.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 
 namespace Roslynator.Metadata;
 
-public record AnalyzerOptionMetadata(string Id, string Key, string DefaultValue, string DefaultValuePlaceholder, string Description)
+public record AnalyzerOptionMetadata(string Id, string Key, string DefaultValue, string DefaultValuePlaceholder, string Description, bool IsObsolete)
 {
     public List<AnalyzerOptionValueMetadata> Values { get; } = new();
 }